% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_lambda.R
\name{bestFitLambda}
\alias{bestFitLambda}
\title{Find best fitted lambda}
\usage{
bestFitLambda(objFit)
}
\arguments{
\item{objFit}{prediction object}
}
\value{
list of predicted lambda for spectra and time
}
\description{
Find best fitted lambda
}
\author{
Asmita Poddar & Florent Latimier
}
