% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{diagCol}
\alias{diagCol}
\title{Diagonal column in parsimonious model}
\usage{
diagCol(data)
}
\arguments{
\item{data}{the data with matrix observation}
}
\description{
Return the list of covariance in column for each clusther in case of a parsimonious model with diagonal column
}
