% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{diagRow}
\alias{diagRow}
\title{Diagonal rows in parsimonious model}
\usage{
diagRow(data)
}
\arguments{
\item{data}{the data with matrix observation}
}
\description{
Return the list of covariance in row for each clusther in case of a parsimonious model with diagonal row
}
