% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpectra.R
\name{fit}
\alias{fit}
\title{Create a list with covariance matrices of the spectra and times}
\usage{
fit(m, modelname = "full", spectra = "diag", time = "diag",
  kerneltypeSpectra = "exponential", kerneltypeTime = "exponential",
  h = 10)
}
\arguments{
\item{m}{spectroscopic data}

\item{modelname}{name of model to be used for calculating the covariance matrix}

\item{spectra}{type of spectra. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{time}{type of time. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{kerneltypeSpectra}{kernel to be used for covariance matrix of spectra
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{kerneltypeTime}{kernel to be used for covariance matrix of time
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{h}{used for kernel calculation}
}
\value{
A list with the covariance matrices for spectra and time, modelname, spectra,
time, weight and mean
}
\description{
Return the covariance matrices
}
\examples{
fittedCov = fit(x, "full")

}
\author{
Asmita Poddar & Florent Latimier
}
