% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpectra.R
\docType{class}
\name{fitData-class}
\alias{fitData-class}
\title{Create a list with covariance matrices of a data with matrix observation.}
\description{
Return the covariance matrices
}
\section{Slots}{

\describe{
\item{\code{m}}{the 3 dimentional data}

\item{\code{modelname}}{name of model to be used for calculating the covariance matrix. Available models are
"full", "parsimonious". Default is "full".}

\item{\code{rows}}{type of rows. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{\code{column}}{type of column. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{\code{kerneltypeRow}}{kernel to be used for covariance matrix of rows
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{\code{kerneltypeCol}}{kernel to be used for covariance matrix of column
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{\code{h}}{used for kernel calculation}

\item{\code{s}}{correction limit paramater for flip flop algorithm}

\item{\code{lambdaR}}{regularisation for rows for flip flop algorithm}

\item{\code{lambdaC}}{regularisation for rows for flip flop algorithm}

\item{\code{validation}}{to optimize lambda in case of th model is : M = parsimonious, S=unknown, T=unknow}

\item{\code{listLambdaR}}{list of lambdaR used in prediction in case validation is TRUE}

\item{\code{listLambdaC}}{list of lambdaC used in prediction in case validation is TRUE}

\item{\code{model}}{use in prediction in case of validation is TRUE}

\item{\code{covMat}}{returning the covariance matrx}
}}

\author{
Asmita Poddar & Florent Latimier
}
