% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitDataMatrix}
\alias{fitDataMatrix}
\title{Fit a 3 dimentional data}
\usage{
fitDataMatrix(m = list(0), modelname = "full", rows = "diag",
  column = "diag", kerneltypeRow = "exponential",
  kerneltypeCol = "exponential", h = 10, s = 0.01, lambdaR = 0.3,
  lambdaC = 0.3, validation = FALSE, listLambdaR = seq(from = 0, to =
  1, by = 0.1), listLambdaC = seq(from = 0, to = 1, by = 0.1),
  model = "gaussian")
}
\arguments{
\item{m}{the 3 dimentional data}

\item{modelname}{name of model to be used for calculating the covariance matrix. Available models are
"full", "parsimonious". Default is "full".}

\item{rows}{type of rows. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{column}{type of column. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{kerneltypeRow}{kernel to be used for covariance matrix of rows
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{kerneltypeCol}{kernel to be used for covariance matrix of column
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{h}{used for kernel calculation}

\item{s}{regularisation paramater for flip flop algorithm}

\item{lambdaR}{regularisation for rows for flip flop algorithm}

\item{lambdaC}{regularisation for rows for flip flop algorithm}

\item{validation}{to optimize lambda or not}

\item{listLambdaR}{list of lambdaR used in prediction in case validation is TRUE}

\item{listLambdaC}{list of lambdaC used in prediction in case validation is TRUE}

\item{model}{use in prediction in case of validation is TRUE}

\item{.Object}{object of class fitData}
}
\value{
a list with covariance matrices for each clusther, of a data with matrix observation
}
\description{
Fit the mean and the covariance matrix according to the model
}
