% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{fitPredPrecent}
\alias{fitPredPrecent}
\title{Predict the label classes of the data}
\usage{
fitPredPrecent(data, modelname = "full", spectra = "diag", time = "diag",
  kerneltypeSpectra = "exponential", kerneltypeTime = "exponential",
  h = 10, lambda = 0.3, model = "gaussian", pTrain = 0.1)
}
\arguments{
\item{data}{spectroscopic data}

\item{modelname}{fitted covariance matrix for the data}

\item{spectra}{type of spectra. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{time}{type of time. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{kerneltypeSpectra}{kernel to be used for covariance matrix of spectra
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{kerneltypeTime}{kernel to be used for covariance matrix of time
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{h}{used for kernel calculation}

\item{lambda}{parameter for regularisation}

\item{model}{type of model to be used for prediction of labels
Available models are "gaussian", "fisher". Default is "gaussian".}

\item{pTrain}{percentage of data to be used for training.}
}
\value{
the percentage of accuracy
}
\description{
Simulate one or more Gaussian spectra at regularly sampling time
}
\examples{
p = fitPredPrecent(m)

}
\author{
Asmita Poddar & Florent Latimier
}
