% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-flop.R
\name{flipflop}
\alias{flipflop}
\title{Implementing the flip flop algorithm}
\usage{
flipflop(data, lmean, s, lambdaS, lambdaT)
}
\arguments{
\item{data}{sata sent}

\item{lmean}{list of means}

\item{s}{parameter for regularisation}

\item{lambdaS}{list of lambda for spectra}

\item{lambdaT}{list of lambda for time}
}
\value{
list of predicted sigma for every label
}
\description{
Implementing the flip flop algorithm
}
\author{
Asmita Poddar & Florent Latimier
}
