% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full.R
\name{full}
\alias{full}
\title{Full model for fitting}
\usage{
full(data)
}
\arguments{
\item{data}{the data with matrix observation}
}
\description{
Return the list of covariance for each clusther in case of a full model.
}
