% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{inversion}
\alias{inversion}
\title{Invert a matrix}
\usage{
inversion(mat)
}
\arguments{
\item{mat}{matrix to be inverted}
}
\value{
Inverted matrix
}
\description{
Invert a matrix
}
\examples{
i = inversion(mat)

}
\author{
Asmita Poddar & Florent Latimier
}
