% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{ker}
\alias{ker}
\title{ker}
\usage{
ker(mat, kernelType, h)
}
\arguments{
\item{mat}{the matrix of distancies}

\item{kernelType}{type of kernel to be apply on the matrix}

\item{h}{the wide parameter to fit the kernel}
}
\description{
Apply a kernel type on a matrix of distancies
}
