% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{kernelRow}
\alias{kernelRow}
\title{Kernel for the row covariance matrix}
\usage{
kernelRow(data, kerneltype, h)
}
\arguments{
\item{data}{the 3 dimentional data}

\item{kerneltype}{a string to represente the type of the kernel.
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{h}{the widths to optimize the kernel}
}
\value{
the kernel square matrix
}
\description{
Calculate the kernel matrix according to the column
}
