% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanData.R
\name{meanData}
\alias{meanData}
\title{Mean of a 3 dimentional data}
\usage{
meanData(data)
}
\arguments{
\item{data}{object of class fitData}
}
\value{
all the means according to labels
}
\description{
Calculate the mean for each label, in row and in column.
}
