% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{percent}
\alias{percent}
\title{Calculate accuracy of prediction}
\usage{
percent(list1, list2)
}
\arguments{
\item{list1}{list containing predicted labels}

\item{list2}{list containing actual ground truth labels}
}
\value{
A floating value of the accuracy percentage
}
\description{
Take two lists containing the predicted and actual labels of the data
}
\examples{
p = percent(p, m$labels)

}
\author{
Asmita Poddar & Florent Latimier
}
