% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMean.R
\name{plot.fitData}
\alias{plot.fitData}
\title{Plot of the fitted mean}
\usage{
\method{plot}{fitData}(data, objFitted)
}
\arguments{
\item{data}{3 dimentional data}

\item{objFitted}{object give by the fitting}
}
\description{
Plot all the row means for a label dependind on the columns.
}
