% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,predictClass-method}
\title{Method num.}
\usage{
predict(Object)

\S4method{predict}{predictClass}(Object)
}
\arguments{
\item{Object}{object to be input}
}
\description{
Method num.

Method num.
}
