% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict the label classes of the data}
\usage{
predict(m, fittedCov, lambda = 0.3, model = "gaussian")
}
\arguments{
\item{m}{spectroscopic data}

\item{fittedCov}{fitted covariance matrix for the data}

\item{lambda}{parameter for regularisation}

\item{model}{type of model to be used for prediction of labels
Available models are "gaussian", "fisher". Default is "gaussian".}
}
\value{
A list with the spectra
}
\description{
Simulate one or more Gaussian spectra at regularly sampling time
}
\examples{
p = predict(m, fittedCov)

}
\author{
Asmita Poddar & Florent Latimier
}
