% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictData}
\alias{predictData}
\title{Prediction of labels}
\usage{
predictData(...)
}
\arguments{
\item{m}{the 3 dimentional data}

\item{fittedCov}{fitted covariance matrix for the data}

\item{lambdaR}{parameter for regularisation of row}

\item{lambdaC}{parameter for regularisation of column}

\item{model}{type of model to be used for prediction of labels
Available models are "gaussian", "tstudent". Default is "gaussian".}

\item{validation}{logical to optimize the lambda.}

\item{listLambdaR}{list of parameter for regularisation of row when do validation}

\item{listLambdaC}{list of parameter for regularisation of column when do validation}

\item{predicted_labels}{predicted class labels}

\item{accuracy}{accracy of prediction}
}
\value{
A list with the row
}
\description{
Predict the label classes of the data according to the fitted arguments.
}
