% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSentinelFiles.R
\name{readMultiSentinelRawData}
\alias{readMultiSentinelRawData}
\title{Read sentinel raw data}
\usage{
readMultiSentinelRawData(path = "./data/", file = "raw.hdf5",
  fileMask = "mask.hdf5", size = NULL)
}
\arguments{
\item{path}{path containing the files}

\item{file}{name of the file with the raw data}

\item{fileMask}{name of the file wiht the mask}

\item{size}{the number of sample to extract}
}
\value{
A list with the sentinel raw data
}
\description{
This function read the sentinel raw data files
}
\examples{
## the famous sentinel data set
\dontrun{
multiSentinelSmoothed <- readMultiSentinelRawData()
}

}
\author{
Serge Iovleff
}
