% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSpectra.R
\name{readMultiSentinelSmoothedData}
\alias{readMultiSentinelSmoothedData}
\title{Read sentinel smoothed data}
\usage{
readMultiSentinelSmoothedData(path = "./data/", file = "gp.hdf5",
  size = NULL)
}
\arguments{
\item{path}{the path containing the files}

\item{file}{the name of the file}

\item{size}{the number of sample to extract}
}
\value{
A list with the sentinel data
}
\description{
This function read the sentinel smoothed data file
}
\examples{
## the famous sentinel data set
\dontrun{
multiSentinelSmoothed <- readMultiSentinelSmoothedData()
}

}
\author{
Serge Iovleff
}
