% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularisation.R
\name{regularisation}
\alias{regularisation}
\title{Perform regularisation on covariance matrix}
\usage{
regularisation(fittedCov, lambdaS = 0.3, lambdaT = 0.3)
}
\arguments{
\item{fittedCov}{fitted covariance matrix}

\item{lambdaS}{regularisation parameter for spectra}

\item{lambdaT}{regularisation parameter for time}
}
\value{
A list with regularised covariance matrix
}
\description{
Perform regularisation on covariance matrix
}
\author{
Asmita Poddar & Florent Latimier
}
