% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate,simulateData-method}
\title{Method num.}
\usage{
simulate(Object)

\S4method{simulate}{simulateData}(Object)
}
\arguments{
\item{Object}{object to be input}
}
\description{
Method num.

#' Method num.
}
