% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateKernel}
\alias{simulateKernel}
\title{SimulateKernel}
\usage{
simulateKernel(modelname, kernelRow, kernelCol, column, rows, labels,
  sigmaL, sigmaR, sigmaC, h)
}
\arguments{
\item{modelname}{type of model to be used to build covariance matrix.
Available options are "full" and "parsimonious". Default is "full".}

\item{kernelRow}{type of kernel to be used to simulate  rows. Available options
are "diag", "epanechnikov", "gaussian", "exponential", "uniform", "quadratic"
, "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "gaussian".}

\item{kernelCol}{type of kernel to be used to simulate  columns. Available options
are "diag", "epanechnikov", "gaussian", "exponential", "uniform", "quadratic"
, "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "gaussian".}

\item{column}{columns intervals of the simulation}

\item{rows}{rows intervals of the simulation}

\item{labels}{class labels of the data}

\item{sigmaL}{the variation due to the clusthers}

\item{sigmaR}{the variation due to the rows}

\item{sigmaC}{the variation due to the columns}

\item{h}{the widths parameter to fit the kernel}
}
\description{
Simulate the covariance matrix as a diagonal or a kenrel
}
