% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSpectra.R
\docType{class}
\name{simulateSpectra}
\alias{simulateSpectra}
\alias{simulateSpectra-class}
\alias{simulateSpectra}
\title{Create a list with a simulated data set of spectra}
\usage{
simulateSpectra(...)
}
\arguments{
\item{...}{any paramaters to be input into the function}
}
\description{
Simulate one or more Gaussian spectra at regularly sampling time

Wrapper function simulaSpectra.
}
\section{Slots}{

\describe{
\item{\code{nbPixel}}{number of pixels belonging to class k}

\item{\code{nbCluster}}{number of cluster}

\item{\code{nbSpectrum}}{number of spectra}

\item{\code{simulationType}}{type of simulation. Available options are "gaussian" and
"tstudent". Default is "gaussian".}

\item{\code{modelname}}{type of model to be used to build covariance matrix.
Available options are "full" and "parsimonious". Default is "full".}

\item{\code{kernelSpectra}}{type of kernel to be used to simulate  spectra. Available options
are "diag", "epanechnikov", "gaussian", "exponential", "uniform", "quadratic"
, "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "gaussian".}

\item{\code{kernelTime}}{type of kernel to be used for simulating time. Available options are
"diag", "epanechnikov", "gaussian", "exponential", "uniform", "quadratic",
"circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "gaussian".}

\item{\code{sigma}}{a vector of size nbSpectrum giving the variance level of
the spectrum}

\item{\code{nbSampling}}{number of time intervals of the simulation}

\item{\code{times}}{time intervals of the simulation}

\item{\code{width}}{the width of the kernel to use for "gaussian" simulation. Default is 50.}

\item{\code{gamma}}{degrees of freedom used for simulating "tstudent" distribution of data.
Default is 3.}

\item{\code{labels}}{class labels of the data}

\item{\code{result}}{return a list of simulated data}
}}

\examples{
m = new("simulateSpectra")
res = simulate(m)

}
\author{
Serge Iovleff, Asmita Poddar & Florent Latimier
}
