% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSpectra.R
\name{simulateSpectra}
\alias{simulateSpectra}
\title{Create a list with a simulated data set of spectra}
\usage{
simulateSpectra(nbPixel, nbCluster, nbSpectrum, kernelName = "gaussian",
  nbSampling = 33, sigma = rexp(nbSpectrum), width = 50)
}
\arguments{
\item{nbPixel}{number of pixels belonging to class k}

\item{nbCluster}{number of cluster}

\item{nbSpectrum}{number of spectra}

\item{kernelName}{[\code{string}] with the kernel to use for the covariance matrix.
Available kernels are "gaussian", "tstudent". Default is "gaussian".}

\item{nbSampling}{number of sampling}

\item{sigma}{a vector of size nbSpectrum giving the variance level of
the spectrum}

\item{width}{the width of the kernel to use for Gaussian simulation. Default is 50.
It also signifies the degree of freedom for Student-T simulation.}
}
\value{
A list with the labels, times, spectra, clouds, means and sigma
}
\description{
Simulate one or more Gaussian spectra at regularly sampling time
}
\examples{
m = simulateSpectra(1000,15,10)

}
\author{
Serge Iovleff & Asmita Poddar
}
