% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{tablesModel}
\alias{tablesModel}
\title{Inversion of matrix}
\usage{
tablesModel(data, lkernelS = list(), lkernelT = list(), lhS = list(),
  lhT = list(), listModel = list("gaussian"))
}
\arguments{
\item{data}{matrix to be inverted}

\item{lkernelS}{list of kernels for spectra}

\item{lkernelT}{list of kernels for time}

\item{lhS}{list of widths for spectra}

\item{lhT}{list of widths for spectra}

\item{listModel}{list of models}
}
\value{
list containing the accuaracy for different models
}
\description{
Inversion of matrix
}
