% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{testTrain}
\alias{testTrain}
\title{Split data into training and test set}
\usage{
testTrain(data, pTrain)
}
\arguments{
\item{data}{spectroscopic data}

\item{pTrain}{percentage of data to be used for training}
}
\value{
A list with training and test data
}
\description{
Split data into training and test set depending on percentage set by the user
}
\examples{
t = testTrain(data, 0.2)

}
\author{
Asmita Poddar & Florent Latimier
}
