% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{unknownCol}
\alias{unknownCol}
\title{Unknown column in parsimonious model}
\usage{
unknownCol(data)
}
\arguments{
\item{data}{the data with matrix observation}
}
\description{
Return the list of covariance in column for each clusther in case of a parsimonious model with unknown column and known rows
}
