% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsimonious.R
\name{unknownRow}
\alias{unknownRow}
\title{Unknown row in parsimonious model}
\usage{
unknownRow(data)
}
\arguments{
\item{data}{the data with matrix observation}
}
\description{
Return the list of covariance in row for each clusther in case of a parsimonious model with unknown rows and known column.
}
