% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestLambda.R
\name{bestFitLambda}
\alias{bestFitLambda}
\title{bestFitLambda}
\usage{
bestFitLambda(objFit)
}
\arguments{
\item{objFit}{fit object}
}
\value{
list of covariance matrix with the best lambda for row and column
}
\description{
Find best fitted lambda for row and column
}
\author{
Asmita Poddar & Florent Latimier
}
