% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_lambda.R
\name{bestPredLambda}
\alias{bestPredLambda}
\title{bestPredLambda}
\usage{
bestPredLambda(objPred)
}
\arguments{
\item{objPred}{prediction object}
}
\value{
list of predicted lambda for row and column and the prediction with accuracy
}
\description{
Find best fitted lambda
}
\author{
Asmita Poddar & Florent Latimier
}
