% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_lambda.R
\name{bestPredLambda}
\alias{bestPredLambda}
\title{Find best fitted lambda}
\usage{
bestPredLambda(objPred, listLambdaS = c(0), listLambdaT = c(0))
}
\arguments{
\item{objPred}{prediction object}

\item{listLambdaS}{list of lambda for spectra}

\item{listLambdaT}{list of lambda for time}
}
\value{
list of predicted lambda for spectra and time
}
\description{
Find best fitted lambda
}
\author{
Asmita Poddar & Florent Latimier
}
