% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpectra.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,fitSpectra-method}
\title{Method num.}
\usage{
fit(Object)

\S4method{fit}{fitSpectra}(Object)
}
\arguments{
\item{Object}{object to be input}
}
\description{
Method num.

#' Method num.
}
