% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpectra.R
\docType{class}
\name{fitSpectra}
\alias{fitSpectra}
\alias{fitSpectra-class}
\title{Create a list with covariance matrices of the spectra and times}
\usage{
fitSpectra(...)
}
\arguments{
\item{...}{any paramaters to be input into the function}
}
\description{
Return the covariance matrices
}
\section{Slots}{

\describe{
\item{\code{m}}{spectroscopic data}

\item{\code{modelname}}{name of model to be used for calculating the covariance matrix. Available models are
"full", "parsimonious". Default is "full".}

\item{\code{spectra}}{type of spectra. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{\code{time}}{type of time. Available models are "diag", "unknown" and "kernel".
Default is "diag".}

\item{\code{kerneltypeSpectra}}{kernel to be used for covariance matrix of spectra
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{\code{kerneltypeTime}}{kernel to be used for covariance matrix of time
Available kernels are "epanechnikov", "gaussian", "exponential", "uniform",
"quadratic", "circular", "triangular", "rational quadratic", "inverse multiquadratic".
Default is "exponential".}

\item{\code{h}}{used for kernel calculation}

\item{\code{s}}{correction limit paramater for flip flop algorithm}

\item{\code{lambdaS}}{regularisation for spectra for flip flop algorithm}

\item{\code{lambdaT}}{regularisation for spectra for flip flop algorithm}

\item{\code{validation}}{to optimize lambda in case of th model is : M = parsimonious, S=unknown, T=unknow}

\item{\code{listLambdaS}}{list of lambdaS used in prediction in case validation is TRUE}

\item{\code{listLambdaT}}{list of lambdaT used in prediction in case validation is TRUE}

\item{\code{model}}{use in prediction in case of validation is TRUE}

\item{\code{covMat}}{returning the covariance matrx}
}}

\author{
Asmita Poddar & Florent Latimier
}
