% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip-flop.R
\name{flipflop}
\alias{flipflop}
\title{Flip - Flop}
\usage{
flipflop(data, lmean, s, lambdaR, lambdaC)
}
\arguments{
\item{data}{sata sent}

\item{s}{parameter for regularisation}

\item{lambdaR}{list of lambda for row}

\item{lambdaC}{list of lambda for column}

\item{list}{of means}
}
\value{
list of predicted sigma for every label
}
\description{
Implementing the flip flop algorithm
}
\author{
Asmita Poddar & Florent Latimier
}
