% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_lambda.R
\name{matxMax}
\alias{matxMax}
\title{Find row and column of maximum element
Find row and column of maximum element}
\usage{
matxMax(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
Vector containing the row and column of maximum element
}
\description{
Find row and column of maximum element
Find row and column of maximum element
}
\author{
Asmita Poddar & Florent Latimier
}
