% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{percent}
\alias{percent}
\title{Find percentage of accuracy
Find percentage of accuracy when two lists containig the class labels are provided}
\usage{
percent(list1, list2)
}
\arguments{
\item{list1}{first list}

\item{list2}{second list}
}
\value{
percentage value - numeric
}
\description{
Find percentage of accuracy
Find percentage of accuracy when two lists containig the class labels are provided
}
\author{
Asmita Poddar & Florent Latimier
}
