% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{class}
\name{predictClass}
\alias{predictClass}
\alias{predictClass-class}
\title{Prediction}
\value{
A list with the row
}
\description{
Predict the label classes of the data with 3 dimensions
}
\section{Slots}{

\describe{
\item{\code{m}}{the 3 dimentional data}

\item{\code{fittedCov}}{fitted covariance matrix for the data}

\item{\code{lambdaR}}{parameter for regularisation of row}

\item{\code{lambdaC}}{parameter for regularisation of column}

\item{\code{model}}{type of model to be used for prediction of labels
Available models are "gaussian", "tstudent". Default is "gaussian".}

\item{\code{validation}}{logical to optimize the lambda.}

\item{\code{listLambdaR}}{list of parameter for regularisation of row when do validation}

\item{\code{listLambdaC}}{list of parameter for regularisation of column when do validation}

\item{\code{predicted_labels}}{predicted class labels}

\item{\code{accuracy}}{accracy of prediction}
}}

\author{
Asmita Poddar & Florent Latimier
}
