% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{class}
\name{predictClass}
\alias{predictClass}
\alias{predictClass-class}
\alias{predictClass}
\title{Predict the label classes of the data}
\usage{
predictClass(...)
}
\arguments{
\item{...}{any paramaters to be input into the function}
}
\value{
A list with the spectra
}
\description{
Simulate one or more Gaussian spectra at regularly sampling time

Wrapper function fitSpectra.
}
\section{Slots}{

\describe{
\item{\code{m}}{spectroscopic data}

\item{\code{fittedCov}}{fitted covariance matrix for the data}

\item{\code{lambdaS}}{parameter for regularisation of spectra}

\item{\code{lambdaT}}{parameter for regularisation of time}

\item{\code{model}}{type of model to be used for prediction of labels
Available models are "gaussian", "fisher". Default is "gaussian".}

\item{\code{validation}}{logical to optimize the lambda.}

\item{\code{predicted_labels}}{predicted class labels}

\item{\code{accuracy}}{accracy of prediction}
}}

\author{
Asmita Poddar & Florent Latimier
}
