% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularisation.R
\name{regularisation}
\alias{regularisation}
\title{Perform regularisation on covariance matrix}
\usage{
regularisation(fittedCov, lambda)
}
\arguments{
\item{fittedCov}{covariance matrix}

\item{lambda}{parameter for regularisation}
}
\value{
Regularised covariance matrix
}
\description{
Perform regularisation on covariance matrix based on the model selected
}
\examples{
reg = regularisation(fittedCov, 0.3)


}
\author{
Asmita Poddar & Florent Latimier
}
