% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularisation.R
\name{regularisation}
\alias{regularisation}
\title{Regularisation}
\usage{
regularisation(fittedCov, lambdaR = 0.3, lambdaC = 0.3)
}
\arguments{
\item{fittedCov}{fitted covariance matrix}

\item{lambdaR}{regularisation parameter for row}

\item{lambdaC}{regularisation parameter for column}
}
\value{
A list with regularised covariance matrix
}
\description{
Perform regularisation on covariance matrix
}
\author{
Asmita Poddar & Florent Latimier

#' @import matrixcalc is.positive.definite
}
