% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSpectra.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simulate,simulateSpectra-method}
\title{Method num.}
\usage{
simulate(Object)

\S4method{simulate}{simulateSpectra}(Object)
}
\arguments{
\item{Object}{object to be input}
}
\description{
Method num.

#' Method num.
}
