% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{tablesModel}
\alias{tablesModel}
\title{Table of all model kind}
\usage{
tablesModel(data, lkernelR = list(), lkernelC = list(), lhR = list(),
  lhC = list(), listModel = list("gaussian"))
}
\arguments{
\item{data}{th 3 dimentinal matrix}

\item{lkernelR}{list of kernels for row}

\item{lkernelC}{list of kernels for column}

\item{lhR}{list of widths for row}

\item{lhC}{list of widths for row}

\item{listModel}{list of models}
}
\value{
list containing the accuaracy for different models
}
\description{
Try all the model posible with validation.
}
