% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatisation.R
\name{testTrain}
\alias{testTrain}
\title{Divide the data into training and test sets
Divide the data into training and test sets according to percentage provided by user}
\usage{
testTrain(data, pTrain)
}
\arguments{
\item{data}{the 3 dimentional data}

\item{pTrain}{the percent of training sample}
}
\value{
list containing training and test set
}
\description{
Divide the data into training and test sets
Divide the data into training and test sets according to percentage provided by user
}
